+++ loggers/winston/CHANGELOG.md
@@ -1,4 +1,40 @@
-# elastic/ecs-winston-format Changelog
+# @elastic/ecs-winston-format Changelog
+
+## Unreleased
+
+- BREAKING CHANGE: Conversion of HTTP request and response objects is no longer
+  done by default. One must use the new `convertReqRes: true` formatter option.
+  As well, only the meta keys `req` and `res` will be handled. Before this
+  change the meta keys `req`, `res`, `request`, and `response` would all be
+  handled.
+
+  Before (no longer works):
+
+  ```
+  const logger = winston.createLogger({
+    format: ecsFormat(),
+    // ...
+  })
+
+  http.createServer(function handler (request, response) {
+    // ...
+    logger.info('handled request', { request, response })
+  })
+  ```
+
+  After:
+
+  ```
+  const logger = winston.createLogger({
+    format: ecsFormat({convertReqRes: true}),  // <-- specify convertReqRes option
+    // ...
+  })
+
+  http.createServer(function handler (req, res) {
+    // ...
+    logger.info('handled request', { req, res })  // <-- only `req` and `res` are special
+  })
+  ```
 
 ## v0.3.0